refactor: remove agent action and move content to agent service with jsdocs
