+++ web-app/src/lib/services/agent/service.ts
@@ -1,11 +1,15 @@
 "use server";
 
+import { revalidatePath } from "next/cache";
+
 import {
+  addAgentToAgentList,
   AgentWithJobs,
   getAgentById,
   getAgentByIdWithPricing,
   getHiredAgents,
   prisma,
+  removeAgentFromAgentList,
 } from "@/lib/db";
 import { JobInputsDataSchemaType } from "@/lib/job-input";
 import { Prisma } from "@/prisma/generated/client";
@@ -14,6 +18,36 @@ import {
   fetchAgentInputSchema,
   getAgentPaymentInformation,
 } from "./third-party";
+
+/**
+ * Toggles an agent's presence in a specified list (adds or removes the agent).
+ *
+ * @param {string} agentId - The unique identifier of the agent to toggle
+ * @param {string} listId - The unique identifier of the list to modify
+ * @param {boolean} isBookmarked - If true, removes the agent from the list; if false, adds the agent to the list
+ * @returns {Promise<{success: boolean}>} An object indicating whether the operation was successful
+ * @throws Will log an error to the console if the operation fails
+ */
+export async function toggleAgentInList(
+  agentId: string,
+  listId: string,
+  isBookmarked: boolean,
+): Promise<{ success: boolean }> {
+  try {
+    if (isBookmarked) {
+      await removeAgentFromAgentList(agentId, listId);
+    } else {
+      await addAgentToAgentList(agentId, listId);
+    }
+
+    // Revalidate the app to update the UI
+    revalidatePath("/app");
+    return { success: true };
+  } catch (error) {
+    console.error("Error toggling agent in list", error);
+    return { success: false };
+  }
+}
 
 export async function getHiredAgentsOrderedByLatestJob(
   userId: string,
